backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-core
langchain-google-genai
langchain-text-splitters
langsmith # Only if you actively use LangSmith for tracing in production. Remove if not.
httpx # Explicitly used in github_data.py
python-dotenv # Optional, for local .env loading, but negligible size.

